Fix schema in async csv writer too. Refactor csv writer to share more code between sync/async paths
