+++ source/Sylvan.Data.Csv/Sylvan.Data.Csv.csproj
@@ -3,7 +3,7 @@
 	<PropertyGroup>
 		<TargetFrameworks>net6.0;netcoreapp3.0;netstandard2.1;netstandard2.0</TargetFrameworks>
 		<VersionPrefix>1.2.6</VersionPrefix>
-		<VersionSuffix>b0001</VersionSuffix>
+		<VersionSuffix>b0002</VersionSuffix>
 		<Description>A .NET library for reading and writing delimited CSV data.</Description>
 		<PackageTags>csv;delimited;data;datareader;datawriter;simd</PackageTags>
 		<Nullable>enable</Nullable>
